+++ services/manifest-repo-export-service/pkg/cmd/server.go
@@ -434,7 +434,7 @@ func processEsls(ctx context.Context, repo repository.Repository, dbHandler *db.
 		}
 		repo.Notify().Notify() // Notify git sync status
 
-		err = service.MeasureGitSyncStatus(ctx, ddMetrics, dbHandler)
+		err = repository.MeasureGitSyncStatus(ctx, ddMetrics, dbHandler)
 		if err != nil {
 			logger.FromContext(ctx).Sugar().Warnf("Failed sending git sync status metrics: %v", err)
 		}
